+++ src/models/Settings.php
@@ -10,6 +10,7 @@ class Settings extends Model
 
     // General
     public $editorUserGroup;
+    public $reviewerUserGroups = [];
     public $publisherUserGroup;
     public $editorNotesRequired = false;
     public $publisherNotesRequired = false;
@@ -17,10 +18,10 @@ class Settings extends Model
     // Notifications
     public $editorNotifications = true;
     public $editorNotificationsOptions = [];
+    public $reviewerNotifications = true;
     public $publisherNotifications = true;
     public $selectedPublishers = '*';
 
     // Permissions
     public $enabledSections = '*';
-
 }